requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
sentence-transformers>=2.2.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0+cpu
transformers>=4.30.0
faiss-cpu>=1.7.4
huggingface-hub>=0.19.0